Extract ``_generate_directives()`` from ``Documenter._generate()``
