+++ sphinx/ext/autodoc/_documenters.py
@@ -1,11 +1,7 @@
 from __future__ import annotations
 
-import sys
-from typing import TYPE_CHECKING, NewType, TypeVar
+from typing import TYPE_CHECKING
 
-from docutils.statemachine import StringList
-
-from sphinx.errors import PycodeError
 from sphinx.ext.autodoc._directive_options import (
     annotation_option,
     bool_option,
@@ -16,24 +12,10 @@ from sphinx.ext.autodoc._directive_options import (
     member_order_option,
     members_option,
 )
-from sphinx.ext.autodoc._docstrings import _prepare_docstrings, _process_docstrings
-from sphinx.ext.autodoc._member_finder import _document_members
-from sphinx.ext.autodoc._renderer import _add_content, _directive_header_lines
-from sphinx.ext.autodoc.mock import ismock
-from sphinx.locale import _, __
-from sphinx.pycode import ModuleAnalyzer
-from sphinx.util import inspect, logging
-from sphinx.util.inspect import safe_getattr
-from sphinx.util.typing import restify, stringify_annotation
 
 if TYPE_CHECKING:
-    from collections.abc import Callable, Sequence
-    from typing import Any, ClassVar, Final, Literal, NoReturn
+    from typing import ClassVar
 
-    from sphinx.config import Config
-    from sphinx.environment import BuildEnvironment, _CurrentDocument
-    from sphinx.events import EventManager
-    from sphinx.ext.autodoc._directive_options import _AutoDocumenterOptions
     from sphinx.ext.autodoc._property_types import (
         _AssignStatementProperties,
         _ClassDefProperties,
@@ -42,18 +24,7 @@ if TYPE_CHECKING:
         _ModuleProperties,
         _TypeStatementProperties,
     )
-    from sphinx.ext.autodoc.importer import _AttrGetter
-    from sphinx.util.typing import OptionSpec, _RestifyMode
-
-logger = logging.getLogger('sphinx.ext.autodoc')
-
-
-def _get_render_mode(
-    typehints_format: Literal['fully-qualified', 'short'],
-) -> _RestifyMode:
-    if typehints_format == 'short':
-        return 'smart'
-    return 'fully-qualified-except-typing'
+    from sphinx.util.typing import OptionSpec
 
 
 class Documenter:
@@ -84,295 +55,6 @@ class Documenter:
         'no-index-entry': bool_option,
         'noindex': bool_option,
     }
-
-    def __init__(
-        self,
-        *,
-        env: BuildEnvironment,
-        options: _AutoDocumenterOptions,
-        get_attr: _AttrGetter,
-        record_dependencies: set[str] | None = None,
-        result: StringList | None = None,
-        indent: str = '',
-    ) -> None:
-        self.config: Config = env.config
-        self.env: BuildEnvironment = env
-        self._current_document: _CurrentDocument = env.current_document
-        self._events: EventManager = env.events
-        self.options: _AutoDocumenterOptions = options
-        self.get_attr = get_attr
-        if record_dependencies is not None:
-            self.record_dependencies: set[str] = record_dependencies
-        else:
-            self.record_dependencies = set()
-        if result is not None:
-            self.result: StringList = result
-        else:
-            self.result = StringList()
-        self.indent: Final = indent
-        # the module analyzer to get at attribute docs, or None
-        self.analyzer: ModuleAnalyzer | None = None
-
-    def add_line(self, line: str, source: str, *lineno: int, indent: str) -> None:
-        """Append one line of generated reST to the output."""
-        if line.strip():  # not a blank line
-            self.result.append(indent + line, source, *lineno)
-        else:
-            self.result.append('', source, *lineno)
-
-    def add_directive_header(self, *, indent: str) -> None:
-        """Add the directive header and options to the generated content."""
-        domain_name = getattr(self, 'domain', 'py')
-        if self.objtype in {'class', 'exception'} and self.props.doc_as_attr:  # type: ignore[attr-defined]
-            directive_name = 'attribute'
-        else:
-            directive_name = getattr(self, 'directivetype', self.objtype)
-        directive_name = f'{domain_name}:{directive_name}'
-
-        if self.analyzer:
-            is_final = self.props.dotted_parts in self.analyzer.finals
-        else:
-            is_final = False
-
-        result = self.result
-        analyzer_source = '' if self.analyzer is None else self.analyzer.srcname
-        source_name = _docstring_source_name(props=self.props, source=analyzer_source)
-        for line in _directive_header_lines(
-            autodoc_typehints=self.config.autodoc_typehints,
-            directive_name=directive_name,
-            is_final=is_final,
-            options=self.options,
-            props=self.props,
-        ):
-            if line.strip():  # not a blank line
-                result.append(indent + line, source_name)
-            else:
-                result.append('', source_name)
-
-    def add_content(self, more_content: StringList | None, *, indent: str) -> None:
-        """Add content from docstrings, attribute documentation and user."""
-        # add content from docstrings
-        attr_docs = {} if self.analyzer is None else self.analyzer.attr_docs
-        analyzer_source = '' if self.analyzer is None else self.analyzer.srcname
-        processed_doc = StringList(
-            list(
-                _process_docstrings(
-                    _prepare_docstrings(props=self.props, attr_docs=attr_docs),
-                    events=self._events,
-                    props=self.props,
-                    obj=self.props._obj,
-                    options=self.options,
-                )
-            ),
-            source=_docstring_source_name(props=self.props, source=analyzer_source),
-        )
-        _add_content(
-            processed_doc,
-            result=self.result,
-            indent=indent + '   ',
-        )
-
-        # add additional content (e.g. from document), if present
-        more_content = self._assemble_more_content(
-            more_content=StringList() if more_content is None else more_content,
-            typehints_format=self.config.autodoc_typehints_format,
-            python_display_short_literal_types=self.config.python_display_short_literal_types,
-            props=self.props,
-        )
-        _add_content(
-            more_content,
-            result=self.result,
-            indent=indent + '   ' * (self.props.obj_type != 'module'),
-        )
-
-    @staticmethod
-    def _assemble_more_content(
-        more_content: StringList,
-        *,
-        props: _ItemProperties,
-        typehints_format: Literal['fully-qualified', 'short'],
-        python_display_short_literal_types: bool,
-    ) -> StringList:
-        """Add content from docstrings, attribute documentation and user."""
-        obj = props._obj
-
-        if props.obj_type in {'data', 'attribute'}:
-            mode = _get_render_mode(typehints_format)
-
-            # Support for documenting GenericAliases
-            if inspect.isgenericalias(obj):
-                alias = restify(obj, mode=mode)
-                more_content.append(_('alias of %s') % alias, '')
-                more_content.append('', '')
-            return more_content
-
-        if props.obj_type in {'class', 'exception'}:
-            from sphinx.ext.autodoc._property_types import _ClassDefProperties
-
-            assert isinstance(props, _ClassDefProperties)
-
-            mode = _get_render_mode(typehints_format)
-
-            if isinstance(obj, NewType):
-                supertype = restify(obj.__supertype__, mode=mode)
-                return StringList([_('alias of %s') % supertype, ''], source='')
-
-            if isinstance(obj, TypeVar):
-                short_literals = python_display_short_literal_types
-                attrs = [
-                    repr(obj.__name__),
-                    *(
-                        stringify_annotation(
-                            constraint, mode, short_literals=short_literals
-                        )
-                        for constraint in obj.__constraints__
-                    ),
-                ]
-                if obj.__bound__:
-                    attrs.append(rf'bound=\ {restify(obj.__bound__, mode=mode)}')
-                if obj.__covariant__:
-                    attrs.append('covariant=True')
-                if obj.__contravariant__:
-                    attrs.append('contravariant=True')
-
-                alias = f'TypeVar({", ".join(attrs)})'
-                return StringList([_('alias of %s') % alias, ''], source='')
-
-            if props.doc_as_attr:
-                try:
-                    analyzer = ModuleAnalyzer.for_module(props.module_name)
-                    analyzer.analyze()
-                    key = ('', props.dotted_parts)
-                    no_classvar_doc_comment = key not in analyzer.attr_docs
-                except PycodeError:
-                    no_classvar_doc_comment = True
-
-                if no_classvar_doc_comment:
-                    alias = restify(obj, mode=mode)
-                    return StringList([_('alias of %s') % alias], source='')
-
-            return more_content
-
-        return more_content
-
-    def _generate(
-        self,
-        more_content: StringList | None = None,
-        real_modname: str | None = None,
-        check_module: bool = False,
-        all_members: bool = False,
-    ) -> None:
-        """Generate reST for the object given by *self.props*, and possibly for
-        its members.
-
-        If *more_content* is given, include that content. If *real_modname* is
-        given, use that module name to find attribute docs. If *check_module* is
-        True, only generate if the object is defined in the module name it is
-        imported from. If *all_members* is True, document all members.
-        """
-        if self.props.obj_type in {'class', 'exception'}:
-            # Do not pass real_modname and use the name from the __module__
-            # attribute of the class.
-            # If a class gets imported into the module real_modname
-            # the analyzer won't find the source of the class, if
-            # it looks in real_modname.
-            real_modname = None
-
-        # If there is no real module defined, figure out which to use.
-        # The real module is used in the module analyzer to look up the module
-        # where the attribute documentation would actually be found in.
-        # This is used for situations where you have a module that collects the
-        # functions and classes of internal submodules.
-        guess_modname = self.props._obj___module__ or self.props.module_name
-        self.real_modname: str = real_modname or guess_modname
-
-        # try to also get a source code analyzer for attribute docs
-        try:
-            self.analyzer = ModuleAnalyzer.for_module(self.real_modname)
-            # parse right now, to get PycodeErrors on parsing (results will
-            # be cached anyway)
-            self.analyzer.find_attr_docs()
-        except PycodeError as exc:
-            logger.debug('[autodoc] module analyzer failed: %s', exc)
-            # no source file -- e.g. for builtin and C modules
-            self.analyzer = None
-            # at least add the module source file as a dependency
-            if self.props.module_name:
-                try:
-                    module_spec = sys.modules[self.props.module_name].__spec__
-                except (AttributeError, KeyError):
-                    pass
-                else:
-                    if (
-                        module_spec is not None
-                        and module_spec.has_location
-                        and module_spec.origin
-                    ):
-                        self.record_dependencies.add(module_spec.origin)
-        else:
-            self.record_dependencies.add(self.analyzer.srcname)
-
-        if self.real_modname != guess_modname:
-            # Add module to dependency list if target object is defined in other module.
-            try:
-                srcname = ModuleAnalyzer.for_module(guess_modname).srcname
-                self.record_dependencies.add(srcname)
-            except PycodeError:
-                pass
-
-        has_docstring = any(self.props._docstrings or ())
-        if ismock(self.props._obj) and not has_docstring:
-            logger.warning(
-                __('A mocked object is detected: %r'),
-                self.props.full_name,
-                type='autodoc',
-                subtype='mocked_object',
-            )
-
-        # check __module__ of object (for members not given explicitly)
-        if check_module and not self.options.imported_members:
-            subject = inspect.unpartial(self.props._obj)
-            modname = self.get_attr(subject, '__module__', None)
-            if modname and modname != self.props.module_name:
-                return
-
-        indent = self.indent
-        analyzer_source = '' if self.analyzer is None else self.analyzer.srcname
-        source_name = _docstring_source_name(props=self.props, source=analyzer_source)
-
-        # make sure that the result starts with an empty line.  This is
-        # necessary for some situations where another directive preprocesses
-        # reST and no starting newline is present
-        self.result.append('', source_name)
-
-        # generate the directive header and options, if applicable
-        self.add_directive_header(indent=indent)
-        self.result.append('', source_name)
-
-        # add all content (from docstrings, attribute docs etc.)
-        self.add_content(more_content, indent=indent)
-
-        # document members, if possible
-        analyzer = self.analyzer
-        if analyzer is not None:
-            analyzer.analyze()
-        _document_members(
-            all_members=all_members,
-            analyzer_order=analyzer.tagorder if analyzer is not None else {},
-            attr_docs=analyzer.attr_docs if analyzer is not None else {},
-            config=self.config,
-            current_document=self._current_document,
-            env=self.env,
-            events=self._events,
-            get_attr=self.get_attr,
-            indent=indent,
-            options=self.options,
-            props=self.props,
-            real_modname=self.real_modname,
-            record_dependencies=self.record_dependencies,
-            registry=self.env._registry,
-            result=self.result,
-        )
 
 
 class ModuleDocumenter(Documenter):
@@ -500,49 +182,3 @@ class TypeAliasDocumenter(Documenter):
         'annotation': annotation_option,
         'no-value': bool_option,
     }
-
-
-class _AutodocAttrGetter:
-    """getattr() override for types such as Zope interfaces."""
-
-    _attr_getters: Sequence[tuple[type, Callable[[Any, str, Any], Any]]]
-
-    __slots__ = ('_attr_getters',)
-
-    def __init__(
-        self, attr_getters: dict[type, Callable[[Any, str, Any], Any]], /
-    ) -> None:
-        super().__setattr__('_attr_getters', tuple(attr_getters.items()))
-
-    def __call__(self, obj: Any, name: str, *defargs: Any) -> Any:
-        for typ, func in self._attr_getters:
-            if isinstance(obj, typ):
-                return func(obj, name, *defargs)
-
-        return safe_getattr(obj, name, *defargs)
-
-    def __repr__(self) -> str:
-        return f'_AutodocAttrGetter({dict(self._attr_getters)!r})'
-
-    def __setattr__(self, key: str, value: Any) -> NoReturn:
-        msg = f'{self.__class__.__name__} is immutable'
-        raise AttributeError(msg)
-
-    def __delattr__(self, key: str) -> NoReturn:
-        msg = f'{self.__class__.__name__} is immutable'
-        raise AttributeError(msg)
-
-
-def _docstring_source_name(*, props: _ItemProperties, source: str) -> str:
-    obj_module = inspect.safe_getattr(props._obj, '__module__', None)
-    obj_qualname = inspect.safe_getattr(props._obj, '__qualname__', None)
-    if obj_module and obj_qualname:
-        # Get the correct location of docstring from props._obj
-        # to support inherited methods
-        fullname = f'{obj_module}.{obj_qualname}'
-    else:
-        fullname = props.full_name
-
-    if source:
-        return f'{source}:docstring of {fullname}'
-    return f'docstring of {fullname}'
