+++ sphinx/ext/autodoc/directive.py
@@ -1,6 +1,5 @@
 from __future__ import annotations
 
-from collections.abc import Callable
 from typing import TYPE_CHECKING
 
 from docutils import nodes
@@ -10,24 +9,56 @@ from sphinx.ext.autodoc._directive_options import (
     _AutoDocumenterOptions,
     _process_documenter_options,
 )
-from sphinx.ext.autodoc._documenters import _AutodocAttrGetter
+from sphinx.ext.autodoc._generate import _generate_directives
 from sphinx.ext.autodoc.importer import _load_object_by_name
 from sphinx.util import logging
 from sphinx.util.docutils import SphinxDirective, switch_source_input
+from sphinx.util.inspect import safe_getattr
 from sphinx.util.parsing import nested_parse_to_nodes
 
 if TYPE_CHECKING:
+    from collections.abc import Callable, Sequence
+    from typing import Any, NoReturn
+
     from docutils.nodes import Node
     from docutils.parsers.rst.states import RSTState
 
-    from sphinx.config import Config
-    from sphinx.ext.autodoc import Documenter
-    from sphinx.ext.autodoc._directive_options import Options
 
 logger = logging.getLogger(__name__)
 
 
-class DummyOptionSpec(dict[str, Callable[[str], str]]):  # NoQA: FURB189
+class _AutodocAttrGetter:
+    """getattr() override for types such as Zope interfaces."""
+
+    _attr_getters: Sequence[tuple[type, Callable[[Any, str, Any], Any]]]
+
+    __slots__ = ('_attr_getters',)
+
+    def __init__(
+        self, attr_getters: dict[type, Callable[[Any, str, Any], Any]], /
+    ) -> None:
+        super().__setattr__('_attr_getters', tuple(attr_getters.items()))
+
+    def __call__(self, obj: Any, name: str, *defargs: Any) -> Any:
+        for typ, func in self._attr_getters:
+            if isinstance(obj, typ):
+                return func(obj, name, *defargs)
+
+        return safe_getattr(obj, name, *defargs)
+
+    def __repr__(self) -> str:
+        return f'_AutodocAttrGetter({dict(self._attr_getters)!r})'
+
+    def __setattr__(self, key: str, value: Any) -> NoReturn:
+        msg = f'{self.__class__.__name__} is immutable'
+        raise AttributeError(msg)
+
+    def __delattr__(self, key: str) -> NoReturn:
+        msg = f'{self.__class__.__name__} is immutable'
+        raise AttributeError(msg)
+
+
+class DummyOptionSpec(dict[str, 'Callable[[str], str]']):  # NoQA: FURB189
     """An option_spec allows any options."""
 
     def __bool__(self) -> bool:
@@ -38,25 +69,12 @@ class DummyOptionSpec(dict[str, Callable[[str], str]]):  # NoQA: FURB189
         return lambda x: x
 
 
-def process_documenter_options(
-    documenter: type[Documenter], config: Config, options: dict[str, str | None]
-) -> Options:
-    from sphinx.ext.autodoc._directive_options import Options
-
-    opts = _process_documenter_options(
-        option_spec=documenter.option_spec,
-        default_options=config.autodoc_default_options,
-        options=options,
-    )
-    return Options(opts)
-
-
 def parse_generated_content(
-    state: RSTState, content: StringList, documenter: Documenter
+    state: RSTState, content: StringList, titles_allowed: bool
 ) -> list[Node]:
     """Parse an item of content generated by Documenter."""
     with switch_source_input(state, content):
-        if documenter.titles_allowed:
+        if titles_allowed:
             return nested_parse_to_nodes(state, content)
 
         node = nodes.paragraph()
@@ -90,14 +108,18 @@ class AutodocDirective(SphinxDirective):
             source, lineno = (None, None)
         logger.debug('[autodoc] %s:%s: input:\n%s', source, lineno, self.block_text)
 
+        registry = self.env._registry
+
         # look up target Documenter
         objtype = self.name[4:]  # strip prefix (auto-).
-        doccls = self.env._registry.documenters[objtype]
+        doccls = registry.documenters[objtype]
+        option_spec = doccls.option_spec
+        titles_allowed = doccls.titles_allowed
 
         # process the options with the selected documenter's option_spec
         try:
             opts = _process_documenter_options(
-                option_spec=doccls.option_spec,
+                option_spec=option_spec,
                 default_options=self.config.autodoc_default_options,
                 options=self.options,
             )
@@ -114,16 +136,22 @@ class AutodocDirective(SphinxDirective):
         documenter_options._tab_width = self.state.document.settings.tab_width
 
         # generate the output
-        get_attr = _AutodocAttrGetter(self.env._registry.autodoc_attrgetters)
+        get_attr = _AutodocAttrGetter(registry.autodoc_attrgetters)
+        name = self.arguments[0]
+        env = self.env
+        config = env.config
+        current_document = env.current_document
+        events = env.events
+
         props = _load_object_by_name(
-            name=self.arguments[0],
+            name=name,
             objtype=objtype,  # type: ignore[arg-type]
-            mock_imports=self.config.autodoc_mock_imports,
-            type_aliases=self.config.autodoc_type_aliases,
-            current_document=self.env.current_document,
-            config=self.config,
-            env=self.env,
-            events=self.env.events,
+            mock_imports=config.autodoc_mock_imports,
+            type_aliases=config.autodoc_type_aliases,
+            current_document=current_document,
+            config=config,
+            env=env,
+            events=events,
             get_attr=get_attr,
             options=documenter_options,
         )
@@ -132,15 +160,19 @@ class AutodocDirective(SphinxDirective):
 
         record_dependencies: set[str] = set()
         result = StringList()
-        documenter = doccls(
-            env=self.env,
-            options=documenter_options,
+        _generate_directives(
+            more_content=self.content,
+            config=config,
+            current_document=current_document,
+            env=env,
+            events=events,
             get_attr=get_attr,
+            indent='',
+            options=documenter_options,
+            props=props,
             record_dependencies=record_dependencies,
             result=result,
         )
-        documenter.props = props
-        documenter._generate(more_content=self.content)
         if not result:
             return []
 
@@ -151,4 +183,4 @@ class AutodocDirective(SphinxDirective):
         for fn in record_dependencies:
             self.state.document.settings.record_dependencies.add(fn)
 
-        return parse_generated_content(self.state, result, documenter)
+        return parse_generated_content(self.state, result, titles_allowed)
